+++ src/main/java/crosis47/minecraft/condense/CondensePlugin.java
@@ -40,7 +40,7 @@ import java.util.Map;
 
 public final class CondensePlugin extends JavaPlugin {
 
-    private static final int CURRENT_CONFIG_VERSION = 7;
+    private static final int CURRENT_CONFIG_VERSION = 8;
     private static final String DEFAULT_CONDENSER_NAME = "Condenser";
 
     private final Set<Material> disabledCondenseInputs = new HashSet<>();